pyproject.toml
[project]
name = "yoni-patent-project"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Eli Young", email = "eliyoung4now@gmail.com" }
]
dependencies = [
    "python-docx>=1.1.2",
    "pypdf2>=3.0.1",
    "pdfplumber>=0.11.4",
    "fitz>=0.0.1.dev2",
    "pymupdf>=1.24.11",
    "pytesseract>=0.3.13",
    "pillow>=10.4.0",
    "pdf2docx>=0.5.8",
    "pypdf>=5.0.1",
    "ocrmypdf>=16.5.0",
    "tesseract>=0.1.3",
    "easyocr>=1.7.2",
    "pypandoc>=1.13",
    "pandoc>=2.4",
    "openai>=1.51.0",
    "python-dotenv>=1.0.1",
    "pyinstaller>=6.10.0",
    "cx-freeze>=7.2.3",
    "tkinterdnd2>=0.4.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []
excluded-dependencies = ["pathlib"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/yoni_patent_project"]
